requirements.txt
attrs==21.4.0
bibtexparser==1.2.0
build==0.7.0
colorama==0.4.4
coverage==6.3.2
distlib==0.3.4
filelock==3.6.0
future==0.18.2
iniconfig==1.1.1
mypy==0.941
mypy-extensions==0.4.3
packaging==21.3
pep517==0.12.0
platformdirs==2.5.1
pluggy==1.0.0
py==1.11.0
pycutils==0.0.8
pyparsing==3.0.6
pytest==7.1.1
pytest-cov==3.0.0
six==1.16.0
toml==0.10.2
tomli==2.0.1
tox==3.24.5
typing_extensions==4.1.1
virtualenv==20.13.4
